--- front/src/components/UI/Button/index.js
@@ -1,1 +1,0 @@
-export { default as Button } from './Button';
