--- front/src/components/UI/CookingTime/CookingTime.module.scss
@@ -1,13 +1,0 @@
-@use 'src/styles/color-variables.scss' as colors;
-
-.cookingTime {
-  background-image: url('assets/images/timer.svg');
-  background-position: left center;
-  background-repeat: no-repeat;
-
-  padding-left: 27px;
-  line-height: 24px;
-
-  color: colors.$color-secondary;
-  white-space: nowrap;
-}
