--- front/src/components/UI/ErrorMessage/ErrorMessage.module.scss
@@ -1,9 +1,0 @@
-@use 'src/styles/variables' as vars;
-
-.error {
-  text-align: center;
-  font-family: vars.$font-body;
-  font-size: vars.$font-size-m;
-  max-width: 500px;
-  margin: 20px auto;
-}
