--- front/src/components/UI/ScrollUpButton/ScrollUpButton.module.scss
@@ -1,17 +1,0 @@
-@use 'src/styles/color-variables' as var;
-
-.btn {
-  position: sticky;
-  right: 0;
-  bottom: 5rem;
-
-  display: flex;
-  margin-left: auto;
-  margin-right: -3.8rem;
-
-  padding: 0;
-  width: 3rem;
-  height: 3rem;
-  border-radius: 50%;
-  box-shadow: 0px 2px 10px 1px var.$shadow-color;
-}
