--- front/src/components/UI/index.js
@@ -1,7 +1,0 @@
-export { ErrorMessage } from './ErrorMessage/ErrorMessage';
-
-export { default as Loader } from './Loader/Loader';
-
-export { CookingTime } from './CookingTime/CookingTime';
-
-export { OvenTime } from './OvenTime/OvenTime';
