+++ front/src/shared/ui/Dropdown/DropdownItem/DropdownItem.module.scss
@@ -1,0 +1,31 @@
+@use 'styles/variables' as vars;
+
+.dropdownItem {
+  box-sizing: border-box;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  height: 40px;
+  padding: 0 16px;
+  border-radius: 8px;
+  background: vars.$bg-primary;
+  font-family: vars.$font-body;
+  font-style: normal;
+  font-weight: 300;
+  font-size: vars.$font-size-xxs;
+  line-height: vars.$line-height-xxs;
+  color: vars.$color-primary;
+
+  &:hover,
+  &:focus,
+  &_active {
+    outline: none;
+    background: vars.$btn-bg-secondary-hover;
+    border-radius: 0;
+  }
+}
+
+.listItem {
+  padding: 0;
+  display: flex;
+}
