+++ front/src/shared/ui/Dropdown/DropdownMenu/DropdownMenu.module.scss
@@ -1,0 +1,57 @@
+@use 'styles/variables' as vars;
+
+.dropdownMenu {
+  position: relative;
+
+  &_type_measure-unit {
+    grid-area: measureUnits;
+  }
+}
+
+.input {
+  box-sizing: border-box;
+  width: 144px;
+  height: 40px;
+  border: 1px solid vars.$input-border-color;
+  border-radius: 8px;
+  padding: 8px 16px;
+  margin: 0;
+  font-family: vars.$font-body;
+  font-style: normal;
+  font-size: vars.$font-size-xxs;
+  font-weight: 500;
+  line-height: vars.$line-height-xxs;
+  color: vars.$color-primary;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  cursor: pointer;
+}
+
+.options {
+  display: none;
+  overflow-y: scroll;
+  position: absolute;
+  z-index: 5;
+  box-sizing: border-box;
+  width: 100%;
+  padding: 8px 0;
+  background: vars.$bg-primary;
+  border-radius: 8px;
+  box-shadow: 0px 4px 12px 5px rgba(1, 7, 62, 0.08);
+  max-height: 340px;
+}
+
+.visible {
+  display: block;
+}
+
+@media screen and (max-width: 576px) {
+  .input {
+    width: 144px;
+  }
+
+  .options {
+    max-height: 176px;
+  }
+}
