+++ front/src/shared/ui/Loader/Loader.jsx
