+++ front/src/shared/ui/PhotoButton/PhotoButton.module.scss
@@ -1,0 +1,78 @@
+@use 'styles/variables' as vars;
+
+.photo {
+  border-radius: 12px;
+
+  &__input {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 20px 0;
+    width: 100%;
+    min-height: 104px;
+    font-family: vars.$font-body;
+    font-size: vars.$font-size-xxs;
+    font-weight: 700;
+    line-height: vars.$line-height-xxs;
+    color: #818181;
+    background-color: #f6f6f6;
+    box-sizing: border-box;
+    border-radius: 12px;
+    transition: border 0.8s;
+    cursor: pointer;
+    border: 1px solid #f6f6f6;
+
+    &:hover,
+    &_on-focus {
+      border: 1px solid #dbdbdb;
+    }
+
+    &_error {
+      border: 1px solid #d14941;
+    }
+
+    &_hidden {
+      display: none;
+    }
+  }
+}
+
+.preview {
+  width: 132px;
+  margin: 0 auto;
+
+  &_visible {
+    display: block;
+  }
+
+  &__container {
+    position: relative;
+    width: 100%;
+    min-height: 132px;
+  }
+
+  &__img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+
+  &__remove {
+    position: absolute;
+    top: 0;
+    right: 0;
+    width: 20px;
+    height: 20px;
+    background-color: transparent;
+    background-image: url('assets/images/remove-img.svg');
+    border: none;
+    opacity: 0.6;
+    transition: opacity 0.6s;
+    cursor: pointer;
+
+    &:hover {
+      opacity: 1;
+    }
+  }
+}
