+++ front/src/shared/ui/StarRating/StarRating.jsx
