+++ front/src/shared/ui/StarRating/index.jsx
