+++ front/src/shared/ui/index.js
@@ -1,0 +1,9 @@
+export { ErrorMessage } from './ErrorMessage/ErrorMessage';
+
+export { default as Loader } from './Loader/Loader';
+
+export { CookingTime } from './CookingTime/CookingTime';
+
+export { OvenTime } from './OvenTime/OvenTime';
+
+export { default as Button } from './Button';
